requested changes
Solved requested changes:
- Corrected typo in  "GridInteractComponent.cpp"
- Translated italian comments in "GridProcessorComponent.cpp"
- "ESurfaceOrientation" and "ETrapType" now derive from uint8

+++ AIV_2025_VGP/Source/AIV_2025_VGP/Public/GridSystem/GridGeneratorVolume.h
@@ -15,7 +15,7 @@
 
 struct FGridSurface;
 enum EObjectTypeQuery;
-enum ESurfaceOrientation : int;
+enum ESurfaceOrientation : uint8;
 enum ECollisionChannel;
 /**
  *
